move stats serialization tests to kernel tests
